test: refactor graph initialization
Use the graph initialization util function in the "serialization" tests.
